core: Do not use channel to order messages for UI assignment
Incoming message handlers need to make sure that generated messages
get UI assigned in the same order as corresponding incoming messages.
Otherwise a replica could assign UIs to its generated messages in
reverse order relative to the corresponding incoming messages. This
could confuse other replicas because they will always process messages
in the order of UI assigned. But they also have to process any
messages extracted from the incoming messages. Moreover, messages
generated by a replica include the corresponding incoming message,
directly or indirectly. Thus, we need to ensure that the order in
which UI gets assigned to the generated messages is consistent with
the order of the corresponding incoming messages.

A first step to achieve this would be to include the process of
assigning UI to the generated message and adding it to the log
directly into the incoming message handler itself rather than doing
that in a separate goroutine. This will also make it easier to
implement generation of ViewChange message later because ViewChange
message should include all previously generated messages with UI. So
it would need to stop processing of incoming messages and wait until
all the message handlers in progress are finished and any generated
messages have got UI assigned and added to the message log. Then it
will be easy to pick up all those messages from the log.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>