core: Use peerstate package to handle replica UIs
A new abstract interface uiReleaser is introduced, and uiAcceptor is
changed to uiCapturer to be able to express two-stage UI handling
model. Both uiReleaser and uiCapturer interface implementations use
peerstate package. Prepare and Commit message handlers use the
supplied uiCapturer and uiReleaser to synchronize beginning and
completion of message processing. This allows messages with UI from
the same peer replica to be processed in sequence eliminating possible
race conditions.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>