usig/sgx: Allow to retrieve USIG epoch value directly
Currently there is no way to get the epoch value out of a SGX USIG
instance without creating a new UI, but that have a side effect of
increasing its counter value. This functionality is going to be useful
for composing a full USIG identity which is USIG public key combined
with the epoch value of the particular USIG instance.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>