core: Remove handling of USIG epoch value from core
MinBFT protocol does not actually define USIG epoch value. The notion
of the epoch value is specific to current SGX USIG implementation and
is used to determine full SGX USIG instance identity. These details
should not be handled by the core of MinBFT protocol. Move this to the
sample implementation of authentication external interface.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>