sample/authentication: Rename SGX USIG authentication scheme
Current implementation of USIG authentication scheme does only support
SGX USIG. This is not easy to make generic because USIG identity
passed to VerifyUI is going to be highly depend on particular USIG
implementation. Current implementation uses serialized public key as
USIG identity. This does not reflect actual SGX USIG identity and is
going to change.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>