core: Change handling of locally generated messages with UI
Messages generated locally with UI attached can be received as part of
some other message from peer replicas. The embedded messages should
normally be extracted and handled using the same generic message
handlers as used to handle incoming messages. When handling a message
produced by the replica itself, there is no need to process the
message, since it must have already been processed when it was
generated. There is also no need to check the UI counter value.
However, the UI attached to the message has to be verified. If the UI
is not valid, then the enclosing message referring to the invalid
message is also deemed invalid and should not be processed.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>