api: Allow state machine operation to return result asynchronously
Return a channel from RequestConsumer.Deliver as a promise to provide
the result of operation execution. Certain types of state machines
could benefit from this by allowing partially concurrent execution of
ordered operations. The consensus protocol does not really need to
wait for the operation to fully complete before supplying the next one
for execution by the state machine.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>