core: Always accept client request identifier
A Request message from a client can arrive to a backup replica
delayed, after it and some subsequent requests have been received from
peer replicas embedded in Prepare and Commit messages, accepted and
executed. This is not necessarily client's fault. The request message
handler should not return error in such case. An identifier of such a
delayed request can be accepted, but simply assumed it is not new,
thus it should not be handled. This will allow to simply restart the
request timer in a backup replica on each new accepted request
identifier per client. This also eliminates any difference in handling
of the Request message received from a client directly or one
extracted from a Prepare message.

On the other hand, a replica can also receive Commit messages delayed,
after the next client Request message has arrived and its request
identifier has been accepted. Thus, it should not be an error trying
to add a Reply message produced as a result of late request execution.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>