core: Allow to prepare/commit requests in advance
There is no need to wait for a request to get executed before
preparing/committing the next one. The primary could receive Commit
messages delayed, but it may generate a new Prepare message for any
subsequent valid Request message as soon as its identifier is greater
than the one in the last generated Prepare message. Similarly, a
backup replica may generate a new Commit message for any new valid
Prepare message, without waiting for any request execution.

This can be achieved by two-stage handling of the client request
identifiers, similarly to UI handling. Firstly, a request identifier
is captured to check if it is new and should be processed. Then it is
processed. The primary generated a new Prepare message, assigns a UI
to it, and appends it to the message log for delivery to the peer
replicas. The backup will re-/start the request timer for the client
at this point. Afterwards, the request identifier is released allowing
to capture and process further Request messages from the corresponding
client. When enough matching valid Commit messages are collected from
different replicas, the request is executed and a Reply message
generated. The Reply message is added to the client state
representation and delivered to the client.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>