core: Ensure exactly-once semantic of request execution
Current implementation of normal mode already prevents the same
request to be executed multiple times by keeping track of prepared
requests. But when the view change is implemented, this mechanism
would not be sufficient. Any single Commit message found in view
change certificate would be enough to safely execute the corresponding
request. The Commit message is not guaranteed to be generated by a
correct replica in that case. Therefore, there has to be an additional
mechanism to ensure exactly-once semantic of request execution. Each
replica needs to keep track of already executed request identifiers
per client and prevent any attempt to execute it again.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>