core/internal/clientstate: Add request timeout handling
Extend client state representation with a module to maintain request
timeout timer. Two new methods allow to re-/start and stop the timer.
Timer implementation and request timeout are to be supplied as
parameters during initialization of the client state representation.
The request timeout is supplied as a function because it may be
required to adjust the timeout during view change operation to
compensate for increasing network delay.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>