core: Implement abstractions to handle request timeout
New abstractions, to get the request timeout, start and stop a request
timer, as well as to handle its expiration, are introduced. A request
timer is to be started in each backup replica whenever a new client
request is received. The timer is to be stopped when the client
request is ready to execute. The expiration handler will trigger the
view change operation, when it is implemented.

When a timeout happens, the timeout handler needs to know what was the
current view number in which the timer was started. This allows to
avoid race conditions and to determine the correct view number to
request to change the view to, given that the handler is invoked
asynchronously.

The standard implementation utilized the client state representation
for maintaining the timer. The abstractions to get request timeout and
to handle the timer expiration are to be used for constructing a new
client state representation.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>