core: Change definition of abstractions to capture/release UI
First, the uiCapturer is changed to receive the message with UI
attached, as opposed to receiving replica ID and parsed UI. Previous
definition was preferred due to the fact that the UI was parsed by
uiVerifer and could then be passed to uiCapturer. Separating message
validation from message processing will not allow to do that anymore.

Second, uiReleaser abstraction is eliminated. The function to release
the captured UI is returned directly from uiCapturer. This allows to
avoid re-parsing the message UI, as well as returning it from
uiCapturer and requiring the caller to pass it back to uiReleaser.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>