messages: Add client ID to Reply message
Reply message lacks client ID field covered by the replica signature.
This opens up possibility of a replay attack. A malicious replica
could theoretically reply to a client with a collection of Reply
messages referring to a request of another client.

Extend Reply message with client ID field covered by the replica
signature so that the client can check if the message was generated in
response to its own Request.

This might also make it possible for a client to store a quorum of
Reply messages for presenting it to a third-party as a proof of
request execution. Another benefit is that this will allow to
better restructure the code in core package.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>