core: Return release function when capturing request ID
As opposed to using a separate function to release a captured request
ID, return such a function when capturing it. This allows to avoid
passing additional arguments and eliminates unnecessary checks in the
release function. Another advantage of this approach is that capturing
and releasing are more tightly coupled in the same scope.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>