core: Generalize collecting commitments on prepared Request
Backup replicas commit to a prepared Request by generating their
Commit messages, whereas the primary does so by the Prepare message
itself. Generalize the abstraction of collecting/counting those
commitments, so as to receive the committing replica identifier
together with the corresponding Prepare message, in place of a Commit
message. This generalization will help to unify message processing by
reusing this abstraction in Prepare processor.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>