CircleCI: Use Pre-Built Docker Image
By this patch, CircleCI uses a pre-built Docker image for CI.

Previously, we introduced a custom image (#50) with the aim of saving
time and resources of a CI process. However, we are faced with problem
that it is difficult to run a CI with a new Dockerfile when it is
updated, as described in #59.

Since we found the saving of the time by a custom image is not so
effective in our project as of now, we will go back to use a pre-built
image.

This closes #59.

Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>