Revert "makefile: Exclude all vendor directories in lint build target"
This reverts commit dfe3cf702ee335f11b6915896013cfac8f3d216a.

We found that the real reason for unexpected reported issues
from gometalinter is mismatch of Go versions used to compile
our project and gometalinter. Refer to the following comment for details.

https://github.com/hyperledger-labs/minbft/pull/58#issuecomment-450424435

In the previous commits, we updated the configuration to solve the mismatch.
So we can revert this workaround.

This closes #51.

Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>