grpc: prevent initialization from failing/blocking
In order to setup a replica group which is tolerant to connectivity
issues, this patch suggests to initialize client stream for each remote
replica in asynchronous and non-blocking manner.  To make this work,
goroutines focusing on sending/receiving messages via the stream need
to wait until the stream is surely initialized.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>