CI: Switch gometalinter to golangci-lint
Gometalinter, a linter used in our CI, will be archived in April 2019.
This patch switches to golangci-lint, another linter introduced as an alternative.

The configuration (linters to be used, skipping test files) are same as before,
except `gotype` is now `typecheck`.
We can find supported linters and which is enabled default in:

gometalinter
https://github.com/alecthomas/gometalinter/blob/v2.0.12/README.md#supported-linters

golangci-lint
https://github.com/golangci/golangci-lint/blob/v1.15.0/README.md#supported-linters

This closes #71.

Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>