README: update requirement for SGX hardware mode
Intel distributes some software components to operate SGX enclaves,
and the required compenents depend on the operation mode.
Currently the README only refered to the requirement for simulation
mode, so let's extend it to support both operation modes.

And we found that it seems not necessary to explicitly write down
the libary path to /etc/ld.so.conf.d/sgx-sdk.conf, so this patch
also removes words about it.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>