README: update installation and setup of Intel SGX SDK
Intel distributes some software components to operate SGX enclaves,
and the required components depend on the operation mode.
Currently the README only refered to the requirement for simulation
mode, so let's extend it to support both operation modes.

An update on /etc/ld.so.conf.d/sgx-sdk.conf is visible byp running
ldconfig, so let's make it clear in the procedure.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>