libusig_shim: change to dynamic linking
Now we don't need to generate "go-wrapper". Before this commit,
libusig_shim had to be built for both SGX hardware mode and simulation
mode individually. The codes of the go-wrapper was generated based on
which builds it depends on. Now they are linked dynamically; the choice
is done not in compile time but runtime.

Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>