api: Make connectivity interface documentation more specific
Be more precise about guarantees provided by ReplicaConnector
interface. Before, "reliable and secure" implied eventual delivery,
integrity, and optional confidentiality. Bounded message delivery
delay is synonymous with eventual delivery and expected in any
practical system; there is no need to particularly expand on that. The
guarantee to preserve the original order of messages was implied by
usage of Go channels, which might not be immediately clear.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>