core: Support view change in commitment counter
After a replica transitions into a new view, it is the new primary's
UI counter values that determine the sequence of requests to be
executed in the new view, and no more requests can be executed in the
old view. Thus, the current state of the commitment counter component
needs to be reset, once the first commitment for the new view gets
accepted. In general, the counter values of different replicas may not
be synchronised, so the component needs to capture the new primary's
counter value from its first commitment, as well.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>