core: Maintain pending request set
A pending request is a request that was processed, but not yet
accepted for execution. The pending request set will be used to
restart consensus after view change is completed. All replicas will
restart timers for all pending requests. In addition, the new primary
will prepare any pending request that does not appear in the new-view
certificate.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>