peer/cmd/request: Introduce request timeout option
MinBFT never fails in theory as long as we meet the requirement of
"<= f faulty nodes", but unfortunately there's no bug-free softwares,
so it's better to prepare for such potential issues to avoid infinite wait.
This patch suggests to add timeout to the sample application.
Note that if timeout value is not specified, sample application waits
forever as it does now.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>