core: Always process embedded messages
Currently, messages embedded in another message are only processed if
the latter will be applied to the replica state. Particularly, if the
message's UI is not the next expected UI from that peer replica or the
message refers to a former view then the embedded messages may not be
processed.

In normal-case, this does not creates any problem, since messages
received directly from correct replicas always come in correct order
and refer to the current view. However, during and after view change,
some messages can be received indirectly, or some delayed but correct
messages may refer to a former view.

Indirectly received messages may not be ready to process until the
embedded messages are processed first. For example, ViewChange message
received indirectly may have a UI that is not the next UI expected by
the receiver from that peer replica, but the message must embed all
the prior messages.

Commit and Prepare messages for former views may embed Request
messages that have not been processed by the receiver. Nevertheless,
such requests can be propagated and executed in a new view. Therefore,
their request identifiers need to be captured first.

Resolve the aforementioned issues by always processing messages
embedded into messages that originate from peer replicas.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>