messages: Use qualifying methods instead of sealed interfaces
The officially recommended way [1] to require types to explicitly
declare that they implement an interface is to add a special method
with a descriptive name to the interface's method set. Then the type
must have an (empty) implementation of that method, in order to
implement the interface. This technique is also known as "qualifying
methods".

[1]: https://golang.org/doc/faq#guarantee_satisfies_interface

Use qualifying methods to ensure the desired semantics of type
switches/assertions for message interfaces.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>