Fix the build status badge in the README
The build status badge in the README is not displayed. This patch fixes
the image source to show the badge.

According to the document, it should be displayed without this patch
but I cannot identify a cause. This patch uses another way (not by a file
path but by a name) to specify the URL of the badge.

[1] https://help.github.com/en/actions/automating-your-workflow-with-github-actions/configuring-a-workflow#adding-a-workflow-status-badge-to-your-repository

Signed-off-by: Yuta Namiki <yuta.namiki@nec.com>