core: Pass view number as parameter to requestApplier
Current implementation of requestApplier performs all the steps that
will be required to restart consensus for a pending request when a
replica completes view change. However, to be used for that purpose,
it should not try acquiring and holding active view, since that must
be done by the caller. To prepare this piece of code for future usage
and simplify further changes, acquire the active view number in
requestProcessor, and pass it to requestApplier as parameter.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>