core: Always process messages for current view
Obeying invariants of the protocol logic, it is safe to take into
account messages referring to the current view, even if the view is
not active, i.e. view change is in progress. This will allow to use
the same logic for processing both messages received during a stable
view and those received after view change was triggered in the
replica, including messages embedded in NewView and ViewChange
messages. However, a replica should skip some actions when processing
messages referring to an inactive current view, e.g. it should not
generate any message for the inactive view.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>