core/internal/viewstate: Do not update expected view in FinishViewChange
This change simplifies the semantics and implementation of the method,
whereas the protocol logic will benefit from starting view change as a
separate step before proceeding to complete it. The rationale is as
follows.

Whenever a replica decides to stop participating in an active view and
begin view change operation, it must make sure other correct replicas
will eventually do the same. Otherwise, other correct replicas might
still continue participating in the old view for arbitrary long time,
and the replica that started view change alone will stop making any
progress. Always starting view change as a separate step will help to
reuse the code that ensures the aforementioned liveness requirement.

Another benefit of always starting view change as a separate step is
that each subsequent view will be terminated with a ViewChange
message. This will make validation of message logs and processing of
new-view certificates easier and less error-prone.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>