introduce SelectiveIgnoreReplicas() to interface api.Configer
Byzantine fault tolerance is core property of MinBFT,
so it's better to test the property in unit test.

This patch suggests to add a new replica configuration
debug.selectiveignorereplicas. This simulates a typical
Byzantine faulty behavior "selective ignore",
For example, consider the case with the following config
in consensus.yaml:

  debug:
    selectiveignorantreplicas:
      0: 1
      1: 0 2

, then replica 0 ignores (i.e. stops sending any message to)
replica 1, and replica 1 should ignore replica 0 and 2.

Signed-off-by: Naoya Horiguchi <naoya.horiguchi@nec.com>