messages: Introduce ViewChange message interface
This commit defines abstract interface for VIEW-CHANGE message type.
Since the checkpoint feature is to be considered later, the interface
omits a method to get the latest stable checkpoint, for simplicity.
Besides that, ViewChange message is augmented with a quorum of
ReqViewChange messages, dubbed "view-change certificate", for the
following reason.

Whenever a replica decides to terminate the current view and start the
view change operation, it must make sure that other correct replicas
will eventually do the same. Otherwise some other correct replica
might stay in the old view for arbitrary long time, and the replica
that started view change would stop making any progress waiting for
other correct replicas.

A replica terminates the current view and generates its ViewChange
message once it has collected a quorum of valid ReqViewChange messages
from different replicas referring to the subsequent view number. Thus,
a simple way to ensure liveness would be to justify the ViewChange
message with a quorum of ReqViewChange messages. This quorum will
ensure view change is also triggered in any correct replica that
processes the ViewChange message.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>