core: Remove operationExecutor
At the cost of introducing an additional abstraction and two atomic
memory operations, operationExecutor only attempts to detect
concurrent execution on the replicated state machine that should be
guaranteed to never happen by the core logic. This seems to be
excessive. Though, such an extra check can be done in the external
replicated state machine component, if desired.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>