core: Retire request identifier in requestExecutor
On one hand, commitmentCollector is overloaded with the actions rather
related committed request execution. On the other hand, the same set
of actions will have to be performed after view change completion,
namely when executing requests propagated to the new view. Therefore,
it is advantageous to move this functionality to requestExecutor that
is dedicated to handle execution of committed requests.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>