core: Apply only messages referring to the current active view
Although, as argued in commit 8e83dc03, it is always safe to take into
account messages referring to the current view, even if the current
view is not active, it is also safe to avoid applying such messages,
because any other correct replica will do the same when it receives
such messages relayed (embedded), since it also starts the view change
before that.

Applying only messages of the current active view appears beneficial,
because checking if the view is active or not when applying messages
complicates the code, whereas replicas anyway need to agree on and
apply the same state when entering a new active view.

Signed-off-by: Sergey Fedorov <sergey.fedorov@neclab.eu>