Update .golangci.yml to disable linter
This change is to forcibly disable some linters to ignore warnings
showw in the recent GitHub Action results.  This is a bad practice
generally speaking, but we do this just to merge "Archive Lab" patch.

Drop some os/Go comination for GitHub Action jobs.

So if someone want to unarchive this project, they first have to reenable
linters and resolve the warnings.

Signed-off-by: Naoya Horiguchi <naoya.horiguchi@nec.com>